revert separate subcomponents
