+++ packages/itwinui-react/src/core/ExpandableBlock/ExpandableBlock.tsx
@@ -1,20 +1,37 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
+import cx from 'classnames';
 import * as React from 'react';
+import {
+  StatusIconMap,
+  SvgChevronRight,
+  Box,
+  useSafeContext,
+  polymorphic,
+  mergeEventHandlers,
+  ButtonBase,
+  useId,
+} from '../../utils/index.js';
 import type { PolymorphicForwardRefComponent } from '../../utils/index.js';
-import type { ExpandableBlockTriggerOwnProps } from './ExpandableBlockTrigger.js';
-import { ExpandableBlockContent } from './ExpandableBlockContent.js';
-import { ExpandableBlockTrigger } from './ExpandableBlockTrigger.js';
-import { ExpandableBlockWrapper } from './ExpandableBlockWrapper.js';
-import { ExpandableBlockEndIcon } from './ExpandableBlockEndIcon.js';
-import { ExpandableBlockCaption } from './ExpandableBlockCaption.js';
-import { ExpandableBlockExpandIcon } from './ExpandableBlockExpandIcon.js';
-import { ExpandableBlockLabelArea } from './ExpandableBlockLabelArea.js';
-import { ExpandableBlockTitle } from './ExpandableBlockTitle.js';
+import { Icon } from '../Icon/Icon.js';
+import { LinkBox } from '../LinkAction/LinkAction.js';
 
-export type ExpandableBlockOwnProps = {
+const ExpandableBlockContext = React.createContext<
+  | ({
+      setExpanded: React.Dispatch<React.SetStateAction<boolean>>;
+      descriptionId: string | undefined;
+      setDescriptionId: React.Dispatch<
+        React.SetStateAction<string | undefined>
+      >;
+    } & ExpandableBlockOwnProps)
+  | undefined
+>(undefined);
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockContext.displayName = 'ExpandableBlockContext';
+}
+
+// ----------------------------------------------------------------------------
+// Main ExpandableBlock component
+
+type ExpandableBlockOwnProps = {
   /**
    * Status of the block.
    * When set, a colored status icon is shown at the end of the main text.
@@ -25,11 +42,6 @@ export type ExpandableBlockOwnProps = {
    * @default false
    */
   isExpanded?: boolean;
-  /**
-   * Whether or not to show the block's content from the start.
-   * @default false
-   */
-  startExpanded?: boolean;
   /**
    * Callback function for toggling an expansion state.
    */
@@ -78,6 +90,234 @@ const ExpandableBlockComponent = React.forwardRef((props, forwardedRef) => {
 >;
 if (process.env.NODE_ENV === 'development') {
   ExpandableBlockComponent.displayName = 'ExpandableBlock';
+}
+
+// ----------------------------------------------------------------------------
+
+const ExpandableBlockWrapper = React.forwardRef((props, forwardedRef) => {
+  const {
+    children,
+    className,
+    onToggle,
+    style,
+    isExpanded,
+    status,
+    size = 'default',
+    styleType = 'default',
+    disabled = false,
+    ...rest
+  } = props;
+
+  const [expandedState, setExpanded] = React.useState(isExpanded ?? false);
+  const expanded = isExpanded ?? expandedState;
+
+  const [descriptionId, setDescriptionId] = React.useState<string | undefined>(
+    undefined,
+  );
+
+  return (
+    <ExpandableBlockContext.Provider
+      value={{
+        status,
+        isExpanded: expanded,
+        onToggle,
+        size,
+        styleType,
+        disabled,
+        setExpanded,
+        children,
+        descriptionId,
+        setDescriptionId,
+      }}
+    >
+      <Box
+        className={cx('iui-expandable-block', className)}
+        data-iui-expanded={expanded}
+        data-iui-size={size}
+        data-iui-variant={styleType !== 'default' ? styleType : undefined}
+        style={style}
+        ref={forwardedRef}
+        {...rest}
+      >
+        {children}
+      </Box>
+    </ExpandableBlockContext.Provider>
+  );
+}) as PolymorphicForwardRefComponent<'div', ExpandableBlockOwnProps>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockWrapper.displayName = 'ExpandableBlock.Wrapper';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.Trigger component
+type ExpandableBlockTriggerOwnProps = {
+  label?: React.ReactNode;
+  caption?: React.ReactNode;
+  expandIcon?: React.ReactNode;
+  endIcon?: React.ReactNode;
+};
+
+const ExpandableBlockTrigger = React.forwardRef((props, forwardedRef) => {
+  const { className, children, label, caption, expandIcon, endIcon, ...rest } =
+    props;
+  const { disabled, status } = useSafeContext(ExpandableBlockContext);
+
+  return (
+    <LinkBox
+      className={cx('iui-expandable-header', className)}
+      data-iui-disabled={disabled ? 'true' : undefined}
+      ref={forwardedRef}
+      {...rest}
+    >
+      {children ?? (
+        <>
+          {expandIcon ?? <ExpandableBlock.ExpandIcon />}
+          <ExpandableBlock.LabelArea>
+            <ExpandableBlock.Title>{label}</ExpandableBlock.Title>
+            {caption && (
+              <ExpandableBlock.Caption>{caption}</ExpandableBlock.Caption>
+            )}
+          </ExpandableBlock.LabelArea>
+          {endIcon || status ? (
+            <ExpandableBlock.EndIcon>{endIcon}</ExpandableBlock.EndIcon>
+          ) : null}
+        </>
+      )}
+    </LinkBox>
+  );
+}) as PolymorphicForwardRefComponent<'div', ExpandableBlockTriggerOwnProps>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockTrigger.displayName = 'ExpandableBlock.Trigger';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.ExpandIcon component
+
+const ExpandableBlockExpandIcon = React.forwardRef((props, forwardedRef) => {
+  const { className, children, ...rest } = props;
+  return (
+    <Icon
+      className={cx('iui-expandable-block-icon', className)}
+      ref={forwardedRef}
+      {...rest}
+    >
+      {children ?? <SvgChevronRight aria-hidden />}
+    </Icon>
+  );
+}) as PolymorphicForwardRefComponent<'span'>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockExpandIcon.displayName = 'ExpandableBlock.ExpandIcon';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.LabelArea component
+
+const ExpandableBlockLabelArea = polymorphic.span('iui-expandable-block-label');
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockLabelArea.displayName = 'ExpandableBlock.LabelArea';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.Title component
+
+const ExpandableBlockTitle = React.forwardRef((props, forwardedRef) => {
+  const { className, children, onClick: onClickProp, ...rest } = props;
+
+  const { isExpanded, setExpanded, disabled, onToggle, descriptionId } =
+    useSafeContext(ExpandableBlockContext);
+
+  return (
+    <ButtonBase
+      className={cx('iui-expandable-block-title', 'iui-link-action', className)}
+      aria-expanded={isExpanded}
+      aria-disabled={disabled}
+      onClick={mergeEventHandlers(onClickProp, () => {
+        if (disabled) {
+          return;
+        }
+        setExpanded(!isExpanded);
+        onToggle?.(!isExpanded);
+      })}
+      ref={forwardedRef}
+      aria-describedby={descriptionId}
+      {...rest}
+    >
+      {children}
+    </ButtonBase>
+  );
+}) as PolymorphicForwardRefComponent<'button'>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockTitle.displayName = 'ExpandableBlock.Title';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.Caption component
+
+const ExpandableBlockCaption = React.forwardRef((props, forwardedRef) => {
+  const fallbackId = useId();
+
+  const { setDescriptionId } = useSafeContext(ExpandableBlockContext);
+
+  React.useEffect(() => {
+    setDescriptionId(props.id || fallbackId);
+    return () => setDescriptionId(undefined);
+  }, [props.id, fallbackId, setDescriptionId]);
+
+  return (
+    <Box
+      ref={forwardedRef}
+      id={fallbackId}
+      {...props}
+      className={cx('iui-expandable-block-caption', props?.className)}
+    />
+  );
+}) as PolymorphicForwardRefComponent<'div'>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockCaption.displayName = 'ExpandableBlock.Caption';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.EndIcon component
+
+const ExpandableBlockEndIcon = React.forwardRef((props, forwardedRef) => {
+  const { children, ...rest } = props;
+  const { status } = useSafeContext(ExpandableBlockContext);
+
+  const icon = children ?? (status && StatusIconMap[status]());
+  return (
+    <Icon fill={status} ref={forwardedRef} {...rest}>
+      {icon}
+    </Icon>
+  );
+}) as PolymorphicForwardRefComponent<
+  'span',
+  React.ComponentPropsWithoutRef<typeof Icon>
+>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockEndIcon.displayName = 'ExpandableBlock.EndIcon';
+}
+
+// ----------------------------------------------------------------------------
+// ExpandableBlock.Content component
+type ExpandableBlockContentOwnProps = {
+  innerProps?: React.ComponentPropsWithoutRef<'div'>;
+};
+
+const ExpandableBlockContent = React.forwardRef((props, forwardedRef) => {
+  const { className, children, innerProps, ...rest } = props;
+
+  return (
+    <Box
+      className={cx('iui-expandable-content', className)}
+      ref={forwardedRef}
+      {...rest}
+    >
+      <Box {...innerProps}>{children}</Box>
+    </Box>
+  );
+}) as PolymorphicForwardRefComponent<'div', ExpandableBlockContentOwnProps>;
+if (process.env.NODE_ENV === 'development') {
+  ExpandableBlockContent.displayName = 'ExpandableBlock.Content';
 }
 
 /**
